+++ backend/api/types.py
@@ -9,7 +9,8 @@ class User(ObjectType):
 class Lobby(ObjectType):
 	code = NonNull(String)
 	mode = NonNull(String)
-	max_rounds = NonNull(String)
+	round_limit = NonNull(Int)
+	max_rounds = NonNull(Int)
 	players = NonNull(List(NonNull(User)))
 	can_start = Boolean()