+++ backend/game/trick.py
@@ -103,7 +103,11 @@ class Trick:
         return after_effects
 
     def __deepcopy__(self, memo):
-        newone = type(self)(None, None, None, None)
-        newone.__dict__.update(self.__dict__)
-        delattr(newone, "history")
-        return newone
+        cls = self.__class__
+        result = cls.__new__(cls)
+        memo[id(self)] = result
+        for k, v in self.__dict__.items():
+            if k == "history":
+                continue
+            setattr(result, k, deepcopy(v, memo))
+        return result