+++ backend/lobby/lobby.py
@@ -1,5 +1,5 @@
-from threading import Lock
-from typing import List, Set
+from threading import RLock
+from typing import Callable, List
 from datetime import datetime
 from copy import deepcopy
 
@@ -10,13 +10,14 @@ from game.card_decks import CardDecks
 
 
 class Lobby:
-	def __init__(self, code: str) -> None:
+	def __init__(self, code: str, on_game_finish: Callable[[GameHistory, "Lobby"], None]) -> None:
 		self.code = code
 		self._settings = Settings("Jubiläum25")
-		self._lock = Lock()
+		self._lock = RLock()
 		self._players: List[User] = []
 		self.created = datetime.now()
 		self._game: Game = None
+		self._on_game_finish = lambda history: on_game_finish(history, self)
 		self.game_started: datetime = None
 
 	def _add_player(self, user: User) -> None:
@@ -72,7 +73,7 @@ class Lobby:
 			# elif len(self._players) < 3:
 			# 	raise Exception("at least 3 players required")
 			self.game_started = datetime.now()
-			self._game = Game(self._players, self._settings)
+			self._game = Game(self._players, self._settings, self._on_game_finish)
 			self._game.setDaemon(True)
 			self._game.start()
 			print("started")
@@ -83,7 +84,7 @@ class Lobby:
 				raise Exception("game not started")
 			return self._game
 
-	def get_game_interaction(self) -> GameHistory:
+	def get_game_history(self) -> GameHistory:
 		with self._lock:
 			if self._game is None:
 				raise Exception("game not started")