+++ frontend/src/theme/styles.ts
@@ -3,7 +3,7 @@ import "./global.css"
 
 export const smoothGradient = (theme: Theme, duration: string) => ({
 	backgroundImage: theme.pretty.primaryGradient,
-	backgroundSize: "200% 200%",
+	backgroundSize: "200% 100%",
 	"&:not(.Mui-disabled)": {
 		animation: `Wizard-shakenNotStirred ${duration} ease infinite`
 	}